Merge pull request #51 from AsadbekRahimov/master

+++ database/migrations/2022_07_17_201242_create_group_rooms_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->integer('group_id');
             $table->integer('room_id');
-            $table->integer('time');
+            $table->float('time');
             $table->timestamps();
         });
     }
